tests fixed

+++ backend/src/database/repositories/memberRepository.ts
@@ -879,18 +879,18 @@ class MemberRepository {
         SELECT
         a."memberId",
         a."segmentId",
-        count(a.id) AS "activityCount",
+        count(a.id)::integer AS "activityCount",
         max(a.timestamp) AS "lastActive",
         array_agg(DISTINCT concat(a.platform, ':', a.type)) FILTER (WHERE a.platform IS NOT NULL) AS "activityTypes",
         array_agg(DISTINCT a.platform) FILTER (WHERE a.platform IS NOT NULL) AS "activeOn",
-        count(DISTINCT a."timestamp"::date) AS "activeDaysCount",
+        count(DISTINCT a."timestamp"::date)::integer AS "activeDaysCount",
         round(avg(
             CASE WHEN (a.sentiment ->> 'sentiment'::text) IS NOT NULL THEN
                 (a.sentiment ->> 'sentiment'::text)::double precision
             ELSE
                 NULL::double precision
             END
-        )::numeric, 2) AS "averageSentiment"
+        )::numeric, 2):: float AS "averageSentiment"
         FROM activities a
         WHERE a."memberId" = :memberId
         GROUP BY a."memberId", a."segmentId"
